Rename 'components/AgencyBarChart.vue' to 'components/index/CardsReference/Agency/Chart.vue'
