+++ components/index/CardsReference/AgencyCard.vue
@@ -1,7 +1,7 @@
 <template>
   <v-col cols="12" md="6" class="DataCard">
     <client-only>
-      <agency-bar-chart
+      <agency-chart
         :title="$t('都庁来庁者数の推移')"
         :title-id="'agency'"
         :chart-id="'agency'"
@@ -12,19 +12,19 @@
         <template v-slot:description>
           {{ $t('※土・日・祝日を除く庁舎開庁日の1週間累計数') }}
         </template>
-      </agency-bar-chart>
+      </agency-chart>
     </client-only>
   </v-col>
 </template>
 
 <script>
-import AgencyBarChart from '@/components/AgencyBarChart.vue'
+import AgencyChart from '@/components/index/CardsReference/Agency/Chart.vue'
 import AgencyData from '@/data/agency.json'
 import { getComplementedDate } from '@/utils/formatDate'
 
 export default {
   components: {
-    AgencyBarChart,
+    AgencyChart,
   },
   data() {
     const labels = AgencyData.labels.map((l) => {
